feat: fix circular dependency on trajectory utils file and divide in multiple files to make it smaller and cleaner
