+++ apps/bilan-carbone/src/components/pages/TrajectoryPage.tsx
@@ -5,7 +5,7 @@ import type { FullStudy } from '@/db/study'
 import { customRich } from '@/i18n/customRich'
 import { getStudyTotalCo2Emissions } from '@/services/study'
 import type { ActionWithRelations, TrajectoryWithObjectivesAndScope } from '@/types/trajectory.types'
-import { getDefaultSnbcSectoralPercentages, getDefaultSnbcSectoralTrajectory } from '@/utils/trajectory'
+import { getDefaultSnbcSectoralPercentages, getDefaultSnbcSectoralTrajectory } from '@/utils/snbc'
 import AddIcon from '@mui/icons-material/Add'
 import { Tooltip, Typography } from '@mui/material'
 import type { ExternalStudy, SectenInfo, TransitionPlan } from '@repo/db-common'
