+++ apps/bilan-carbone/src/components/pages/TransitionPlanInitPage.tsx
@@ -9,7 +9,7 @@ import OnboardingSectionStep from '@/components/study/transitionPlan/OnboardingS
 import SectorAllocationBlock from '@/components/study/transitionPlan/SectorAllocationBlock'
 import TrajectoryGraph from '@/components/study/transitionPlan/TrajectoryGraph'
 import TransitionPlanFilters from '@/components/study/transitionPlan/TransitionPlanFilters'
-import { TRAJECTORY_15_ID, TRAJECTORY_SNBC_GENERAL_ID } from '@/constants/trajectories'
+import { TRAJECTORY_15_ID, TRAJECTORY_SNBC_GENERAL_ID } from '@/constants/trajectory.constants'
 import type { FullStudy } from '@/db/study'
 import { useLocalStorageSync } from '@/hooks/useLocalStorageSync'
 import { useServerFunction } from '@/hooks/useServerFunction'
@@ -25,8 +25,7 @@ import {
 } from '@/services/serverFunctions/transitionPlan'
 import { TrajectoryWithObjectivesAndScope } from '@/types/trajectory.types'
 import { compareSectenVersions } from '@/utils/secten'
-import { calculateSectoralSNBCReductionRates } from '@/utils/snbc'
-import { getDefaultSnbcSectoralTrajectory } from '@/utils/trajectory'
+import { calculateSectoralSNBCReductionRates, getDefaultSnbcSectoralTrajectory } from '@/utils/snbc'
 import DeleteIcon from '@mui/icons-material/Delete'
 import type { ExternalStudy, SectenInfo, SectenVersion, TransitionPlan } from '@repo/db-common'
 import { TrajectoryType } from '@repo/db-common/enums'
