+++ apps/bilan-carbone/src/components/study/trajectory/TrajectoryCreationModal.tsx
@@ -2,6 +2,7 @@
 
 import LoadingButton from '@/components/base/LoadingButton'
 import Modal from '@/components/modals/Modal'
+import { SBTI_START_YEAR, SNBC_FINAL_TARGET_YEAR } from '@/constants/trajectory.constants'
 import { useServerFunction } from '@/hooks/useServerFunction'
 import {
   createTrajectorySchema,
@@ -15,19 +16,17 @@ import {
 } from '@/services/serverFunctions/trajectory.serverFunction'
 import type { BaseObjective, PastStudy } from '@/types/trajectory.types'
 import { TrajectoryWithObjectivesAndScope } from '@/types/trajectory.types'
+import { getDefaultSBTIReductionRate } from '@/utils/sbti'
 import {
   calculateBaseSNBCReductionRates,
   calculateSectoralSNBCReductionRates,
   extractSNBCReductionRatesFromObjectives,
-  SNBC_FINAL_TARGET_YEAR,
 } from '@/utils/snbc'
 import { getYearFromDateStr } from '@/utils/time'
 import {
   getCorrectedObjectives,
   getDefaultObjectivesForTrajectoryType,
-  getDefaultSBTIReductionRate,
   getDisplayedReferenceYearForTrajectoryType,
-  SBTI_START_YEAR,
 } from '@/utils/trajectory'
 import { zodResolver } from '@hookform/resolvers/zod'
 import type { SectenInfo } from '@repo/db-common'
