+++ apps/bilan-carbone/src/components/study/trajectory/TrajectoryCreationStep2.tsx
@@ -6,9 +6,9 @@ import GlossaryModal from '@/components/modals/GlossaryModal'
 import { customRich } from '@/i18n/customRich'
 import { TrajectoryFormData } from '@/services/serverFunctions/trajectory.command'
 import type { BaseObjective } from '@/types/trajectory.types'
+import { getDefaultSBTIReductionRate } from '@/utils/sbti'
 import { ReductionRates } from '@/utils/snbc'
 import { toTitleCase } from '@/utils/string'
-import { getDefaultSBTIReductionRate } from '@/utils/trajectory'
 import { FormControlLabel, Radio, RadioGroup, Typography } from '@mui/material'
 import { TrajectoryType } from '@repo/db-common/enums'
 import classNames from 'classnames'
