+++ apps/bilan-carbone/src/constants/trajectory.constants.ts
@@ -20,3 +20,11 @@ export const SNBC_SECTOR_TARGET_EMISSIONS: Record<SectenSector, { 2015: number; 
   [TRAJECTORY_SNBC_AGRICULTURE_ID]: { 2015: 87000, 2030: 73000, 2050: 48000 },
   [TRAJECTORY_SNBC_TRANSPORTATION_ID]: { 2015: 138000, 2030: 99000, 2050: 4000 },
 }
+export const SBTI_REDUCTION_RATE_15 = 0.042
+export const SBTI_REDUCTION_RATE_WB2C = 0.025
+export const SBTI_START_YEAR = 2020
+export const MID_TARGET_YEAR = 2030
+export const TARGET_YEAR = 2050
+export const OVERSHOOT_THRESHOLD = 0.05
+export const SNBC_REFERENCE_YEAR = 1990
+export const SNBC_FINAL_TARGET_YEAR = 2050
