+++ apps/bilan-carbone/src/types/trajectory.types.ts
@@ -85,6 +85,7 @@ export type ActionWithRelations = Action & {
   tags: Array<ActionTag & { studyTag: StudyTag }>
   subPosts: ActionSubPost[]
 }
+
 export interface TrajectoryData {
   previousTrajectoryStartYear: number | null
   previousTrajectory: TrajectoryDataPoint[] | null
@@ -100,6 +101,7 @@ export interface TrajectoryResult {
   customTrajectories: Array<{ id: string; data: TrajectoryData }>
   actionBased: TrajectoryData | null
 }
+
 export type TrajectorySeries = LineSeriesType & {
   dataType: DataType
   isFailed?: boolean
